Separate `set_global_shift_decrease!` and `set_global_shift_increase!` and utilize `m[2]` (take 2) (#59)
set_global_shift! was already a giant if statement with two sides, this splits those sides into two functions (NFC perf, slightly increased length, IMO more clarity and easier reasoning because it is more clear which call sites can lead to which branches). In theory this could improve perf but I don't measure it.

Then, utilizing that increased clarity, depend on m[2] in set_global_shift_decrease! for performance improvement in the pathological2 case.

(take 2, see #56 for the first try)

Adds tests that failed reliably in the first try.