Move changing m[3] to increase m[4] to above 2^32 to _rand_slow_path.
The point of that increase is to make sure we hit the slow path infrequently so we only need to do it if
we are actually hitting the slow path. This reduces the scope of some pathological cases which also makes
them slightly less pathological because they require three operations to trigger rather than two.